# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agenda Inteligente — agenda-service",
...
    "version": "0.4.0"
  },
  "servers": [
    {
      "url": "https://cpdf-agenda.better-knowledge.com",
      "description": "Protótipo do programa"
    }
  ],
  "paths": {
...
    "/waitlist": {
      "post": {
        "tags": [
          "fila de espera"
        ],
        "summary": "Entra na fila de espera para uma janela de horário",
        "description": "**Escopo:** `agenda:write`\n\nUse quando o horário desejado está ocupado (RF-14). A fila é por JANELA ('quinta à tarde'), não por horário exato — para um horário livre, agende direto. Quando um cancelamento liberar um horário compatível, o primeiro da fila recebe a oferta pelo canal. **Não há reserva**: o horário segue livre na grade e quem confirmar primeiro leva. Aceita Idempotency-Key.",
        "operationId": "entrar_waitlist_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WaitlistIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaitlistOut"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "examples": {
                  "NAO_AUTENTICADO": {
                    "summary": "Credencial ausente ou inválida",
                    "value": {
                      "code": "NAO_AUTENTICADO",
                      "message": "Credencial ausente ou inválida: nenhum header de credencial presente",
                      "hint": "Envie `Authorization: Bearer <jwt do Supabase>` ou `X-Agent-Key`.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "403": {
            "description": "Escopo insuficiente",
            "content": {
              "application/json": {
                "examples": {
                  "ESCOPO_INSUFICIENTE": {
                    "summary": "A credencial não tem o escopo exigido",
                    "value": {
                      "code": "ESCOPO_INSUFICIENTE",
                      "message": "A credencial não tem o escopo 'agenda:cancel'.",
                      "hint": "Peça uma credencial com o escopo necessário — cancelamento exige agenda:cancel.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "422": {
            "description": "Validação falhou",
            "content": {
              "application/json": {
                "examples": {
                  "PAYLOAD_INVALIDO": {
                    "summary": "Corpo ou parâmetros não passaram na validação",
                    "value": {
                      "code": "PAYLOAD_INVALIDO",
                      "message": "Payload inválido: body.cliente_telefone: String should have at least 8 characters",
                      "hint": "Corrija os campos apontados e repita a chamada com os mesmos dados.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "404": {
            "description": "Não encontrado nesta organização",
            "content": {
              "application/json": {
                "examples": {
                  "NAO_ENCONTRADO": {
                    "summary": "Id não existe nesta organização",
                    "value": {
                      "code": "NAO_ENCONTRADO",
                      "message": "Serviço 6f1e…9c02 não existe nesta organização.",
                      "hint": "Confira o id — liste o recurso correspondente para obter ids válidos.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "400": {
            "description": "Pedido malformado",
            "content": {
              "application/json": {
                "examples": {
                  "DATA_SEM_FUSO": {
                    "summary": "Horário sem offset de fuso",
                    "value": {
                      "code": "DATA_SEM_FUSO",
                      "message": "'inicio' veio sem offset de fuso horário.",
                      "hint": "Envie ISO 8601 com offset explícito, ex.: 2026-08-27T15:30:00-03:00.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          }
        },
        "x-escopo-requerido": "agenda:write",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Repetir a chamada com a mesma chave devolve a resposta original sem duplicar efeito. Obrigatório para agentes; use um UUID novo por intenção."
          }
        ]
      },
      "get": {
        "tags": [
          "fila de espera"
        ],
        "summary": "Fila de espera da organização, na ordem de chegada",
        "description": "**Escopo:** `agenda:read`\n\nPor padrão mostra quem ainda espera (aguardando/ofertado). A posição é por serviço.",
        "operationId": "listar_waitlist_get",
        "parameters": [
          {
            "name": "service_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Service Id"
            }
          },
          {
            "name": "incluir_encerrados",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Também mostra aceito/expirado/cancelado",
              "default": false,
              "title": "Incluir Encerrados"
            },
            "description": "Também mostra aceito/expirado/cancelado"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WaitlistOut"
                  },
                  "title": "Response Listar Waitlist Get"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "examples": {
                  "NAO_AUTENTICADO": {
                    "summary": "Credencial ausente ou inválida",
                    "value": {
                      "code": "NAO_AUTENTICADO",
                      "message": "Credencial ausente ou inválida: nenhum header de credencial presente",
                      "hint": "Envie `Authorization: Bearer <jwt do Supabase>` ou `X-Agent-Key`.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "403": {
            "description": "Escopo insuficiente",
            "content": {
              "application/json": {
                "examples": {
                  "ESCOPO_INSUFICIENTE": {
                    "summary": "A credencial não tem o escopo exigido",
                    "value": {
                      "code": "ESCOPO_INSUFICIENTE",
                      "message": "A credencial não tem o escopo 'agenda:cancel'.",
                      "hint": "Peça uma credencial com o escopo necessário — cancelamento exige agenda:cancel.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "422": {
            "description": "Validação falhou",
            "content": {
              "application/json": {
                "examples": {
                  "PAYLOAD_INVALIDO": {
                    "summary": "Corpo ou parâmetros não passaram na validação",
                    "value": {
                      "code": "PAYLOAD_INVALIDO",
                      "message": "Payload inválido: body.cliente_telefone: String should have at least 8 characters",
                      "hint": "Corrija os campos apontados e repita a chamada com os mesmos dados.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "502": {
            "description": "Dependência não respondeu — retry é seguro",
            "content": {
              "application/json": {
                "examples": {
                  "CANAL_INDISPONIVEL": {
                    "summary": "O canal-service não respondeu",
                    "value": {
                      "code": "CANAL_INDISPONIVEL",
                      "message": "O canal de WhatsApp não respondeu.",
                      "hint": "Tente de novo em instantes — a operação é segura para repetir.",
                      "retryable": true
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          }
        },
        "x-escopo-requerido": "agenda:read"
      }
    },
    "/waitlist/{entry_id}": {
      "delete": {
        "tags": [
          "fila de espera"
        ],
        "summary": "Sai da fila de espera",
        "description": "**Escopo:** `agenda:write`\n\nIdempotente: sair de novo devolve o mesmo resultado.",
        "operationId": "sair_waitlist__entry_id__delete",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Entry Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaitlistOut"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "examples": {
                  "NAO_AUTENTICADO": {
                    "summary": "Credencial ausente ou inválida",
                    "value": {
                      "code": "NAO_AUTENTICADO",
                      "message": "Credencial ausente ou inválida: nenhum header de credencial presente",
                      "hint": "Envie `Authorization: Bearer <jwt do Supabase>` ou `X-Agent-Key`.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "403": {
            "description": "Escopo insuficiente",
            "content": {
              "application/json": {
                "examples": {
                  "ESCOPO_INSUFICIENTE": {
                    "summary": "A credencial não tem o escopo exigido",
                    "value": {
                      "code": "ESCOPO_INSUFICIENTE",
                      "message": "A credencial não tem o escopo 'agenda:cancel'.",
                      "hint": "Peça uma credencial com o escopo necessário — cancelamento exige agenda:cancel.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "422": {
            "description": "Validação falhou",
            "content": {
              "application/json": {
                "examples": {
                  "PAYLOAD_INVALIDO": {
                    "summary": "Corpo ou parâmetros não passaram na validação",
                    "value": {
                      "code": "PAYLOAD_INVALIDO",
                      "message": "Payload inválido: body.cliente_telefone: String should have at least 8 characters",
                      "hint": "Corrija os campos apontados e repita a chamada com os mesmos dados.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "404": {
            "description": "Não encontrado nesta organização",
            "content": {
              "application/json": {
                "examples": {
                  "NAO_ENCONTRADO": {
                    "summary": "Id não existe nesta organização",
                    "value": {
                      "code": "NAO_ENCONTRADO",
                      "message": "Serviço 6f1e…9c02 não existe nesta organização.",
                      "hint": "Confira o id — liste o recurso correspondente para obter ids válidos.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          }
        },
        "x-escopo-requerido": "agenda:write"
      }
    },
...
    "/canal/optouts": {
      "get": {
        "tags": [
          "canal"
        ],
        "summary": "Clientes que pediram para não receber mensagem ativa",
        "operationId": "listar_optouts_canal_optouts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CanalOptoutOut"
                  },
                  "type": "array",
                  "title": "Response Listar Optouts Canal Optouts Get"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                },
                "examples": {
                  "NAO_AUTENTICADO": {
                    "summary": "Credencial ausente ou inválida",
                    "value": {
                      "code": "NAO_AUTENTICADO",
                      "message": "Credencial ausente ou inválida: nenhum header de credencial presente",
                      "hint": "Envie `Authorization: Bearer <jwt do Supabase>` ou `X-Agent-Key`.",
                      "retryable": false
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Escopo insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                },
                "examples": {
                  "ESCOPO_INSUFICIENTE": {
                    "summary": "A credencial não tem o escopo exigido",
                    "value": {
                      "code": "ESCOPO_INSUFICIENTE",
                      "message": "A credencial não tem o escopo 'agenda:cancel'.",
                      "hint": "Peça uma credencial com o escopo necessário — cancelamento exige agenda:cancel.",
                      "retryable": false
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validação falhou",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                },
                "examples": {
                  "PAYLOAD_INVALIDO": {
                    "summary": "Corpo ou parâmetros não passaram na validação",
                    "value": {
                      "code": "PAYLOAD_INVALIDO",
                      "message": "Payload inválido: body.cliente_telefone: String should have at least 8 characters",
                      "hint": "Corrija os campos apontados e repita a chamada com os mesmos dados.",
                      "retryable": false
                    }
                  }
                }
              }
            }
          },
          "502": {
            "description": "Dependência não respondeu — retry é seguro",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                },
                "examples": {
                  "CANAL_INDISPONIVEL": {
                    "summary": "O canal-service não respondeu",
                    "value": {
                      "code": "CANAL_INDISPONIVEL",
                      "message": "O canal de WhatsApp não respondeu.",
                      "hint": "Tente de novo em instantes — a operação é segura para repetir.",
                      "retryable": true
                    }
                  }
                }
              }
            }
          }
        },
        "x-escopo-requerido": "canal:admin",
        "description": "**Escopo:** `canal:admin`"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CanalOptoutOut": {
        "properties": {
          "telefone": {
            "type": "string",
            "title": "Telefone",
            "description": "Endereço no canal: E.164 ou tg:<chat_id>"
          },
          "origem": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origem",
            "description": "palavra_chave | pedido_humano"
          },
          "em": {
            "type": "string",
            "title": "Em"
          }
        },
        "type": "object",
        "required": [
          "telefone",
          "em"
        ],
        "title": "CanalOptoutOut"
      },
...
      "WaitlistIn": {
        "properties": {
          "service_id": {
            "type": "string",
            "format": "uuid",
            "title": "Service Id"
          },
          "cliente_nome": {
            "type": "string",
            "minLength": 1,
            "title": "Cliente Nome"
          },
          "cliente_telefone": {
            "type": "string",
            "minLength": 3,
            "title": "Cliente Telefone",
            "description": "E.164 no WhatsApp ou tg:<chat_id> no Telegram"
          },
          "janela_inicio": {
            "type": "string",
            "format": "date-time",
            "title": "Janela Inicio",
            "description": "Começo da janela desejada, ISO 8601 com offset"
          },
          "janela_fim": {
            "type": "string",
            "format": "date-time",
            "title": "Janela Fim",
            "description": "Fim da janela desejada, ISO 8601 com offset"
          },
          "resource_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Id",
            "description": "Opcional: só aceita com este profissional/sala"
          }
        },
        "type": "object",
        "required": [
          "service_id",
          "cliente_nome",
          "cliente_telefone",
          "janela_inicio",
          "janela_fim"
        ],
        "title": "WaitlistIn",
        "description": "Entrar na fila para uma JANELA, não para um horário: quem quer um\nhorário específico e livre simplesmente agenda.",
        "examples": [
          {
            "cliente_nome": "Paula Andrade",
            "cliente_telefone": "+5511998765432",
            "janela_fim": "2026-08-27T18:00:00-03:00",
            "janela_inicio": "2026-08-27T12:00:00-03:00",
            "service_id": "b3f0a1d4-2c5e-4f6a-8b7c-9d0e1f2a3b4c"
          }
        ]
      },
      "WaitlistOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "service_id": {
            "type": "string",
            "format": "uuid",
            "title": "Service Id"
          },
          "resource_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Id"
          },
          "cliente_nome": {
            "type": "string",
            "title": "Cliente Nome"
          },
          "cliente_telefone": {
            "type": "string",
            "title": "Cliente Telefone"
          },
          "janela_inicio": {
            "type": "string",
            "format": "date-time",
            "title": "Janela Inicio"
          },
          "janela_fim": {
            "type": "string",
            "format": "date-time",
            "title": "Janela Fim"
          },
          "janela_humana": {
            "type": "string",
            "title": "Janela Humana",
            "description": "A janela desejada em pt-BR, pronta para falar"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "aguardando | ofertado | aceito | expirado | cancelado"
          },
          "posicao": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posicao",
            "description": "Posição na fila entre os que aguardam o mesmo serviço"
          },
          "expira_em": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expira Em",
            "description": "Quando a oferta em aberto expira (status=ofertado)"
          },
          "slot_ofertado": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slot Ofertado",
            "description": "Horário exato proposto pela oferta em aberto"
          },
          "avisos": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Avisos",
            "description": "O que o prestador precisa saber — ex.: cliente em opt-out não recebe oferta"
          }
        },
        "type": "object",
        "required": [
          "id",
          "service_id",
          "resource_id",
          "cliente_nome",
          "cliente_telefone",
          "janela_inicio",
          "janela_fim",
          "janela_humana",
          "status"
        ],
        "title": "WaitlistOut"
      }
    },
    "securitySchemes": {
      "SupabaseJWT": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT do Supabase Auth com claim `org_id` — humanos e UI."
      },
      "AgentKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Agent-Key",
        "description": "Chave estática de agente (fase 1 do conector — vira OAuth 2.1 no agenda-mcp). Ações irreversíveis pedem confirmação humana."
      }
    }
  },
...
  "security": [
    {
      "SupabaseJWT": []
    },
    {
      "AgentKey": []
    }
  ]
}